Add files via upload
The liver is an organ about the size of a football. It sits just under your rib cage on the right side of your abdomen. The liver is essential for digesting food and ridding your body of toxic substances (Read_more).  we build a gradient boosting machine learning algorithm to predict the 4 stages of liver disease in the dataset. The 3 stages of data analytics are Descriptive Analytics, Diagnostics Analytics, and Predictive Analysis. using the methodology of the data science project. The evaluation metric is F1_score and we have an average score of 50.29%.